bot/requirements.txt
tweepy>=4.14,<5
pyyaml>=6.0,<7
